+++ exercises/1901050024/1001S02E03_calculator.py
@@ -1,0 +1,83 @@
+print("Easy Calculator")
+number_1 = float(input('请输入数字：'))
+formula = input('请输入公式+、-、*、/：')
+number_2 = float(input('请输入数字：'))
+if formula == str('+'):
+    print(number_1+number_2)
+elif formula == str('-'):
+    print(number_1-number_2)
+elif formula == str('*'):
+    print(number_1*number_2)
+elif formula == str('/'):
+    print(number_1/number_2)
+else: 
+    print("please enter 1the right formula")
+print("Thank you") 
+
+
+print("简易计算器")
+print ("请输入你想要计算的类型")
+def add(x,y):
+    result = x + y
+    return result
+def sub(x,y):
+    result = x-y
+    return result
+def mul(x,y):
+    result = x*y
+    return result
+def div(x,y):
+    result = x/y
+    return result
+a = input("1:加法；2:减法；3:乘法；4:除法:\n")
+b = float(input('请输入数字:'))
+c = float(input('请输入数字:'))
+if a == '1' :
+        print(b,'+',c,'=',add(b,c))
+elif a == '2':
+        print(b,'-',c,'=',sub(b,c))
+elif a == '3':
+        print(b,'*',c,'=',mul(b,c))
+elif a == '4':
+        print(b,'/',c,'=',div(b,c))
+else: 
+        print("输入错误")
+print("结束")
+
+
+
+def calculate():
+    formula = input('''\
+    Easy Calculator
+    请输入公式+、-、*、/：
+''')
+    number_1 = float(input('请输入数字：'))
+    number_2 = float(input('请输入数字：'))
+    if formula == str('+'):
+            print(number_1+number_2)
+    elif formula == str('-'):
+            print(number_1-number_2)
+    elif formula == str('*'):
+            print(number_1*number_2)
+    elif formula == str('/'):
+            print(number_1/number_2)
+    else: 
+            print("please enter 1the right formula")
+            return calculate()
+    
+        # Add again() function to calculate() function
+    again()
+
+def again():
+    calc_again = input('''
+Do you want to calculate again?
+Please type Y for YES or N fosr NO.
+''')
+
+    if calc_again.upper() == 'Y':
+        calculate()
+    elif calc_again.upper() == 'N':
+        print('See you later.')
+    else:
+        again()
+calculate()
