--- exercises/1901050024/copy3.py
@@ -1,48 +1,0 @@
-def calculate():
-    operation = input('''
-Please type in the math operation you would like to complete:
-+ for addition
-- for subtraction
-* for multiplication
-/ for division
-''')
-
-    number_1 = int(input('Please enter the first number: '))
-    number_2 = int(input('Please enter the second number: '))
-
-    if operation == '+':
-        print('{} + {} = '.format(number_1, number_2))
-        print(number_1 + number_2)
-
-    elif operation == '-':
-        print('{} - {} = '.format(number_1, number_2))
-        print(number_1 - number_2)
-
-    elif operation == '*':
-        print('{} * {} = '.format(number_1, number_2))
-        print(number_1 * number_2)
-
-    elif operation == '/':
-        print('{} / {} = '.format(number_1, number_2))
-        print(number_1 / number_2)
-
-    else:
-        print('You have not typed a valid operator, please run the program again.')
-
-    # Add again() function to calculate() function
-    again()
-
-def again():
-    calc_again = input('''
-Do you want to calculate again?
-Please type Y for YES or N fosr NO.
-''')
-
-    if calc_again.upper() == 'Y':
-        calculate()
-    elif calc_again.upper() == 'N':
-        print('See you later.')
-    else:
-        again()
-
-calculate()
